Dates with a different format were not being edited correctly.
There are two problems - firstly rails doesn't seem to be honouring the Date::DATE_FORMATS[:default] when dates are being converted from strings during update_attributes so the conversion are done manually now.

Secondly the Date to json conversion was also ignoring the default DATE_FORMATS setting so the to string conversions are done before the data is handed off as part of the render operation.